+++ src/lib/flows/common/common-helper.js
@@ -1,15 +1,27 @@
+const { logger } = require("../../common");
 const { execute } = require("../../command/command");
 const { treatCommand } = require("../../command/command-treatment-delegator");
 const { getDir } = require("./build-chain-flow-helper");
 const core = require("@actions/core");
 
 async function executeBuild(rootFolder, nodeChain, projectTriggeringJob) {
-  for await (const node of nodeChain) {
-    await executeNodeBuildCommands(
-      rootFolder,
-      node,
-      projectTriggeringJob === node.project
-    );
+  const projectTriggeringJobIndex = nodeChain.findIndex(
+    node => node.project === projectTriggeringJob
+  );
+  for await (const [index, node] of nodeChain.entries()) {
+    if (node.build && node.build.skip) {
+      logger.info(
+        `Execution skip for ${node.project}. No command will be executed.`
+      );
+    } else {
+      const levelType =
+        index < projectTriggeringJobIndex
+          ? "upstream"
+          : index == projectTriggeringJobIndex
+          ? "current"
+          : "downstream";
+      await executeNodeBuildCommands(rootFolder, node, levelType);
+    }
   }
 }
 
@@ -20,37 +32,37 @@ async function executeBuildSpecificCommand(rootFolder, nodeChain, command) {
   }
 }
 
-async function executeNodeBuildCommands(
-  rootFolder,
-  node,
-  isProjectTriggeringJob
-) {
+/**
+ *
+ * @param {String} rootFolder the folder path to execute command
+ * @param {Object} node the node to execute
+ * @param {String} levelType an option between upstream, current or downstream
+ */
+async function executeNodeBuildCommands(rootFolder, node, levelType) {
   const dir = getDir(rootFolder, node.project);
   if (node.build["build-command"].before) {
     await executeBuildCommands(
       dir,
-      getCommand(node.build["build-command"].before, isProjectTriggeringJob),
+      getCommand(node.build["build-command"].before, levelType),
       node.project
     );
   }
   await executeBuildCommands(
     dir,
-    getCommand(node.build["build-command"], isProjectTriggeringJob),
+    getCommand(node.build["build-command"], levelType),
     node.project
   );
   if (node.build["build-command"].after) {
     await executeBuildCommands(
       dir,
-      getCommand(node.build["build-command"].after, isProjectTriggeringJob),
+      getCommand(node.build["build-command"].after, levelType),
       node.project
     );
   }
 }
 
-function getCommand(buildCommand, isProjectTriggeringJob) {
-  return isProjectTriggeringJob
-    ? buildCommand.current
-    : buildCommand.upstream || buildCommand.current;
+function getCommand(buildCommand, levelType) {
+  return buildCommand[levelType] || buildCommand.current;
 }
 
 async function executeBuildCommands(cwd, buildCommands, project) {